requirements.txt
numpy==1.17.0
labelme==3.16.2
ipdb==0.13.9
pandas==1.1.5
tqdm==4.42.1
matplotlib==3.1.2
torch==1.6.0
torchvision==0.7.0 
albumentations==1.0.3
opencv_python_headless==4.5.3.56
Pillow==8.3.1
scikit_learn==0.24.2
opencv-python==4.1.2.30
opencv-contrib-python==4.1.2.30